+++ CHANGELOG.md
@@ -35,5 +35,6 @@ Component-specific:
 Config-specific:
 - [#42] Implement config repositories
 - [#48] Save config diff in cluster
+- [#39] Encrypt sensitive data
 - [#36] Set registry configuration for dogu and global config.
 - [#45] Set registry configuration for encrypted values.